requirements.txt
# Core dependencies
numpy>=1.21.0
torch>=1.9.0
torchvision>=0.10.0
pillow>=8.3.1
opencv-python>=4.5.3
trimesh>=3.9.0
scipy>=1.7.0

# 3D generation models
git+https://github.com/openai/shap-e.git
# Shap-E dependencies
einops>=0.6.0
transformers>=4.15.0

# Visualization
matplotlib>=3.4.3

# Utilities
tqdm>=4.62.0
argparse>=1.4.0